FB-40681 Disallow updates to existing description updates
Do not allow updates to an existing change of the same description when
it already exists. Previously, uniqueness was defined based upon the
text of the change. As a result, updating a file would cause the change
to rerun even though it may not be safe to do so.

With these changes, updating a change that has already been applied will
have no effect.