FB-40681 Do not update if description changes
The latest change intended to prevent updates from occurring when the
SQL had changed but the description was constant. The intention was to
prevent updates that had already been applied from being re-applied.

However, after that change it was discovered that uptodater files have
been modified in the past to change their description but keep the same
SQL. This use case was not anticipated and as a result our change caused
those uptodaters to be eligible for reapplication. This is a problem.

The code has thus been updated to handle both possible cases of
previously known uptodaters:
- Same SQL
- Same description

In either case, the uptodater should not be eligible for reapplication.

This commit handles these two cases. This handles the following cases:
- Already run uptodater description renamed (as was discovered)
- Already run uptodater SQL changed (what we're trying to prevent)